refactored, but errs with promises
